refactor: domyland chat module name
